legal-ai-layer2/requirements.txt

fastapi
uvicorn
python-multipart
streamlit
google-generativeai
requests
beautifulsoup4
PyPDF2
python-docx
pillow
pytesseract
python-jose
fastapi
uvicorn
pydantic
google-genai
requests
ollama
ddgs
rapidfuzz
langdetect
deep_translator
gtts
pdfplumber
pymongo
email-validator
passlib
bcrypt
passlib[bcrypt]
openai-whisper
langchain
faiss-cpu
